chore(api): separate tempRoute into module
